Refactor endpoint engine config helpers
